+++ src/UserInterface/JsonRpcQueueItemHandlerFactory.php
@@ -37,6 +37,8 @@ final class JsonRpcQueueItemHandlerFactory implements JsonRpcQueueItemHandlerFac
             'exit'                                    => 'exitJsonRpcQueueItemHandler',
             'shutdown'                                => 'shutdownJsonRpcQueueItemHandler',
             'workspace/didChangeWatchedFiles'         => 'didChangeWatchedFilesJsonRpcQueueItemHandler',
+            'textDocument/didOpen'                    => 'didOpenJsonRpcQueueItemHandler',
+            'textDocument/didClose'                   => 'didCloseJsonRpcQueueItemHandler',
             'textDocument/didChange'                  => 'didChangeJsonRpcQueueItemHandler',
             'textDocument/didSave'                    => 'didSaveJsonRpcQueueItemHandler',
             'textDocument/completion'                 => 'completionJsonRpcQueueItemHandler',