+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ## 5.0.0-RC3 (Unreleased)
 * [Fix `uris` in project files being ignored and not analyzed](https://gitlab.com/Serenata/Serenata/issues/243)
 * [Fix crash when analyzing erroneous code containing multiple object arrows, such as `$test->->`](https://gitlab.com/Serenata/Serenata/issues/245)
+* [Stub `textDocument/didOpen`](https://gitlab.com/Serenata/Serenata/issues/253)
+* [Stub `textDocument/didClose`](https://gitlab.com/Serenata/Serenata/issues/259)
 
 ## 5.0.0-RC2
 * Port fixes performed in version 4.3.1